Pre-Restructuring Commit
Modified the set of notebooks so that there is now a notebook for initializing the analysis that includes compiling data from the various files associated with a single daaRun. Following this commit, I'm removing code from downstream notebooks to compile data and will instead rely on opening the existing files to load data and then generate plots.