Pre-major change commit
I realized that the 0.0 initialization notebook won't work as planned because the whole point is that it generates a threshold dataframe for manual, unblinded threshold entry before the user performs manual waveform analysis. So now I need to change the code to make this spreadsheet without relying on the analyzed file.